[Refactor] UI optimaztion and modularization
