+++ lib/widgets/custom_dropdown.dart
@@ -1,0 +1,42 @@
+import 'package:flutter/material.dart';
+
+class CustomDropdown extends StatelessWidget {
+  final String? value;
+  final String hint;
+  final List<String> items;
+  final ValueChanged<String?> onChanged;
+
+  const CustomDropdown({
+    Key? key,
+    required this.value,
+    required this.hint,
+    required this.items,
+    required this.onChanged,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return DropdownButtonFormField<String>(
+      value: value ?? 'All',
+      decoration: InputDecoration(
+        hintText: hint,
+        border: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(10),
+        ),
+        contentPadding: const EdgeInsets.symmetric(horizontal: 15, vertical: 10),
+        filled: true,
+        fillColor: Colors.white,
+      ),
+      items: items
+          .map((item) => DropdownMenuItem(
+                value: item,
+                child: Text(
+                  item,
+                  style: const TextStyle(color: Colors.black),
+                ),
+              ))
+          .toList(),
+      onChanged: onChanged,
+    );
+  }
+}
